v1

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pakopian <akopyanpyotrx99@gmail.com>       +#+  +:+       +#+        */
+/*   By: ggevorgi <sp1tak.gg@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/18 21:28:48 by pakopian          #+#    #+#             */
-/*   Updated: 2025/03/26 21:23:23 by pakopian         ###   ########.fr       */
+/*   Updated: 2025/03/29 19:55:48 by ggevorgi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,13 @@
 # include <stdarg.h>
 # include <unistd.h>
 
-int ft_putchar(char c);
-int ft_printf(const char *str, ...);
-int print_string(char *str);
-void print_hex(unsigned int num, char spec, int *res);
-void print_int(int n, int *res);
-void print_unsigned(unsigned int n, int *res);
-void print_pointer(unsigned long pt, int *res);
-int ft_strlen(char *str);
+int		ft_putchar(char c);
+int		ft_printf(const char *str, ...);
+int		print_string(char *str);
+void	print_hex(unsigned long long num, char spec, int *res);
+void	print_int(int n, int *res);
+void	print_unsigned(unsigned int n, int *res);
+void	print_pointer(unsigned long long pt, int *res);
+int		ft_strlen(char *str);
 
 #endif
